added quests_killcount.txt
This new file will be used to know (or guess at least) the kill count
goal for any known quest. This was created because some servers never
send this info, or at least not when we want ite :)

The file will be updated when new info is sent from server.

Needs more testing.