Changes to misc inventory hook calls
Added item object to hook args. (If we remove it from inventoryList then
how would we get the item based on the index? (didn't think about it
when I made the hook))
Changed $item->{invIndex} to $invIndex to guarantee that we send the
right former index even if in the future inventoryList::remove changes
$item->{invIndex}.
Added a check to amount arg, if for some reason we got a negative amount
just send 0.