account_related branch - make Account module!
Implementing the plan to pull account related data and routines out of
the many scattered modules of Openkore, and into it's own module:
Account.pm

The 'plan' and inspiration for this is found in our wiki:

http://openkore.com/index.php/Project/Refactoring#Account-related

This initial commit establishes this branch (account_related) and
the creation of the Account module.   It is not yet functional.

Following commits in this branch will begin the move of variables out
of Globals.pm that belong in Account.pm

Added new file
new file:   src/Account.pm