Committed change
I added a very quick tutorial on how to run on cmd. I personally think that we should at least show them how to git clone, rather then just mentioning the method. This would prevent users who could not download with the link provided to raise an issue because they can not get both methods to work. (ie. download manager problems)

It would further lead to confusion when the user tries to create an issue ticket because they could not understand what are "commit" and "branch" from the following sentence
``` Before submitting an issue, ensure you are running the latest commit on branch. ```

Users might decide to leave this tool because it is too complicated.

Hope I proved PR. Otherwise, I'll be more than happy to improve it.
Thanks!